added ComesBefore to Polygon, some notes, and some sample (non-functional) code to MyPolygons)
Created the implementation of ComesBefore(), which will take a new polygon and compare its area and distance with the current polygon, returning either true or false depending on whether the original polygon or the new one is supposed to come first (true) or the parsed polygon (false).

Also added notes pertaining to my lack of understanding of interfaces x3.

also added a small amount of code to MyPolygons, very basically using the ComesBefore method.